requirements.txt
autopep8==1.4.4
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Django==2.2.5
django-background-tasks==1.2.0
django-compat==1.0.15
djangorestframework==3.10.3
html5lib==1.0.1
idna==2.8
lxml==4.4.1
pkg-resources==0.0.0
pycodestyle==2.5.0
pytz==2019.2
requests==2.22.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
urllib3==1.25.3
webencodings==0.5.1
